# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This a toolset server.",
        "title": "ToolSet API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/loader/shellcode-runner": {
            "post": {
                "description": "Generate Runner according to shellcode provided by the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream",
                    "application/json"
                ],
                "tags": [
                    "loader"
                ],
                "summary": "generate a shellcode runner",
                "parameters": [
                    {
                        "format": "base64",
                        "description": "the param to generate shellcode runner",
                        "name": "runner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ShellcodeRunner"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "shellcode runner",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "model.ShellcodeRunner": {
            "type": "object",
            "properties": {
                "regen": {
                    "description": "ReGen Whether to regenerate the loader, if true, it will remove cache, this may be beneficial for bypass AV",
                    "type": "boolean",
                    "example": true
                },
                "shellcode": {
                    "description": "Shellcode This is a base64 encoded shellcode",
                    "type": "string",
                    "example": "MTIzemN4"
                },
                "x64": {
                    "description": "X64 Whether the shellcode is x64",
                    "type": "boolean",
                    "example": false
                }
            }
        }
    }
}